requirements.txt
# requirements.txt
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2+cu117
pydicom==2.4.4
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0
matplotlib==3.8.3
flask==2.3.3
gunicorn==22.0.0
albumentations==1.4.0
huggingface_hub==0.20.3
scikit-learn==1.4.1.post1
torchmetrics==0.11.4
tqdm==4.66.2
pandas==2.2.2
datasets==2.16.1
tensorboard==2.15.1
seaborn==0.13.0